refactor : voice 도메인을 post 외부로 분리
