+++ src/main/java/inu/codin/codin/domain/voice/repository/VoiceRepository.java
@@ -1,7 +1,7 @@
-package inu.codin.codin.domain.post.domain.voice.repository;
+package inu.codin.codin.domain.voice.repository;
 
 import inu.codin.codin.common.dto.Department;
-import inu.codin.codin.domain.post.domain.voice.entity.VoiceEntity;
+import inu.codin.codin.domain.voice.entity.VoiceEntity;
 import org.bson.types.ObjectId;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -9,7 +9,6 @@ import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.data.mongodb.repository.Query;
 import org.springframework.stereotype.Repository;
 
-import java.util.List;
 import java.util.Optional;
 
 @Repository
